Smoothly raise the stack to visualize time between new rows
Similar to Tetris Attack, we draw the stack slightly higher as the bump
timer decreases. This makes the adding of new rows far less abrupt.
However, manually bumping the stack is still instant.

Also, the raising of the stack is paused when blocks are cleared,
regardless of the time until the next bump.