Implement 2-speed cursor; Allow multi-directional input
Fixes #12

When holding a directional button long enough, the block cursor will
have a lower cooldown time on its movement. This makes it easier to get
from one side of the playing field to the other.

In addition, it is also now possible to use multiple directional keys at
once, allowing the player to move diagonally.